refactor(ios): extract sub-views from BottomInfoSheet.swift (#A1)
BottomInfoSheet.swift was 1818 lines, well over the 800-line project
limit. Extracted three files:
- NearbyExperienceRow.swift (516 lines) — card row for each experience
- NearbySectionView.swift (406 lines) — NearbySection, skeleton, empty state
- RoutesSectionView.swift (89 lines) — RoutesSection, RouteNowSectionHeader

BottomInfoSheet.swift is now 818 lines. No behavioral changes.
